Added execute_trajectory_em with metadata Moved Flyer class definition to 63-
